Added cornered resize support
Each corner now comes with its own handle that can change the size of the bounding box in 2 dimensions at the same time. (Note the cursor does not actually change to a diagonal arrow to indicate when hovering over a corner. No idea how to get that feature but we should probably look into that)