added column to track time user spent on a label
Added a column to track the amount of time it took for a user to make a label. That time is currently defined as
- (time of frist save  - time between opening page ) + time between frist and second save + ..... + time between n -1 and nth save.

If user leaves page and returns back to the label then it adds on the time so
previosus_time_spent + (time of opening page - time of frist save after returning to label) + etc.

Also added a check to see if segment is already saved so it does not waste time saving a previously saved clip (and messing up the time spent count)).

Also when going to an audio clip with previously exsiting segmentations, those segmentations will appear as saved since they are already in the database