Hotfix 1 (#92)
* added column to track time user spent on a label

Added a column to track the amount of time it took for a user to make a label. That time is currently defined as
- (time of frist save  - time between opening page ) + time between frist and second save + ..... + time between n -1 and nth save.

If user leaves page and returns back to the label then it adds on the time so
previosus_time_spent + (time of opening page - time of frist save after returning to label) + etc.

Also added a check to see if segment is already saved so it does not waste time saving a previously saved clip (and messing up the time spent count)).

Also when going to an audio clip with previously exsiting segmentations, those segmentations will appear as saved since they are already in the database

* Update JsonLabelsToCsv.py

Added time_spent to annotation csv

* Take datewise MySQL backups (#89)

* Take datewise MySQL backups

* Added a fix to combine working mysql start up code with rclone setup

* Linux combability fixes

Tested on my linux machine. I've made changes on both linux and windows for these changes. Basically made sure the system knew were the entrypoint script was.

* Update docker-compose.prod.yml

* Enable installing cron and fix crontab

* Remove conainer startup command from docker compose

* Added more fixes to make sure it works on production server

* Move cron startup to mysql docker entrypoint file

* Remove redundant custom startup script

* Dump mysql databases as a single file

* Include hour and minute information in mysql file name

Co-authored-by: Sean1572 <sean.hyatt.perry@gmail.com>
Co-authored-by: root <root@e4e-acoustics.sdsc.edu>

* Update crontab

Changed cron to run every 6th hour at the top of the hour for production

Co-authored-by: Vaibhav Kant Tiwari <vktiwari@ucsd.edu>
Co-authored-by: root <root@e4e-acoustics.sdsc.edu>