Turned regions into bounding boxes -(wait for feature toggle till merge) (#80)
* Part one: rendering handles

Got the bottom and top handles to render, still need to make them actually change shape

* Top bounding box side resize

Added ability to resize the top part of the region. Still need to:
- add same feature to the bottom
-move box up and down
- Do a 2d scale when creating a region in the frist place

and that is just the frontend features

* Bounding on bottom

Bounding box now can be adjusted via top and bottom, now I just need to make it drag up and down

* Adding 2d dragging!

You can now drag a region across the spectrogram and it looks cool.

Next step would be to do 2d click and drag to create a region and set both x and y bounds (currently only creates in y direction)

* added init regions in 2d!

When creating a region, you can init the bounding box on both x and y values.

Next steps
- Convert height values of y axis into frequency measurements
- Takes those measurements and added them to backend when saving the regions
- Display frequency measurements on data download files

Optional -

* Added cornered resize support

Each corner now comes with its own handle that can change the size of the bounding box in 2 dimensions at the same time. (Note the cursor does not actually change to a diagonal arrow to indicate when hovering over a corner. No idea how to get that feature but we should probably look into that)

* Added instance vars to track actual freqency values

Added instance variables to track the frequency values and added them to annotate_c.js to send to backend

Next step is creating the tables needed to store that data (which means adding more columns to the database)

* Added json output for frequencies

Frequecies are picked up on json. Next step is getting it into a csv

* CSV Frequency data

CSV added frequency data to it, but I forgot about going back to data (and now it looks big). Will go back to fix that

* Added y-axis segmentation genreation when reviewing previously done clips

When going back to completed audio files, you can now see the fully boxed clip.

* Fixed rebase errors

* linting

* Added toggleable feature + More Linting

* ran lint

* Added 2d Labels, feature form intergration